+++ routes/index.js
@@ -1,5 +1,4 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 
 const auth = require('../middlewares/auth');
 const userRouter = require('./users');
@@ -7,29 +6,16 @@ const movieRouter = require('./movies');
 const NotFoundError = require('../errors/not-found-error');
 
 const { login, createUser } = require('../controllers/users');
+const { validateAuthorization, validateRegistration } = require('../middlewares/validation');
 
-router.post('/signup', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-    name: Joi.string().required().min(2).max(30),
-  }),
-}), createUser);
+router.post('/signup', validateRegistration, createUser);
 
-router.post('/signin', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-  }),
-}), login);
+router.post('/signin', validateAuthorization, login);
 
 router.use(auth);
 
 router.use('/users', userRouter);
 router.use('/movies', movieRouter);
-
-//router.use('/', require('./users'));
-//router.use('/', require('./movies'));
 
 router.use('*', () => {
   throw new NotFoundError('Страница не найдена.');
