+++ routes/users.js
@@ -1,15 +1,9 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 
 const { getUser, updateUser } = require('../controllers/users');
+const { validateUpdateUser } = require('../middlewares/validation');
 
 router.get('/me', getUser);
-
-router.patch('/me', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    name: Joi.string().required().min(2).max(30),
-  }),
-}), updateUser);
+router.patch('/me', validateUpdateUser, updateUser);
 
 module.exports = router;
